添加rt_ak.ai_free()
