+++ project_0/.config
@@ -566,7 +566,7 @@ CONFIG_PKG_MICROPYTHON_PATH="/packages/language/micropython"
 # CONFIG_MICROPYTHON_USING_MACHINE_WDT is not set
 # CONFIG_MICROPYTHON_USING_MACHINE_TIMER is not set
 # CONFIG_MICROPYTHON_USING_NETWORK is not set
-# CONFIG_MICROPYTHON_USING_MACHINE_LCD is not set
+CONFIG_MICROPYTHON_USING_MACHINE_LCD=y
 # end of Hardware Module
 
 #
@@ -1037,7 +1037,19 @@ CONFIG_BSP_LCD_Y_MAX=320
 # CONFIG_BSP_USING_PWM is not set
 # CONFIG_BSP_USING_TIMER0 is not set
 # CONFIG_BSP_USING_TIMER1 is not set
-# CONFIG_BSP_USING_CAMERA is not set
+CONFIG_BSP_USING_CAMERA=y
+CONFIG_CAMERA_USING_GC0308=y
+# CONFIG_CAMERA_USING_OV2640 is not set
+# CONFIG_CAMERA_USING_OTHER is not set
+CONFIG_BSP_CAMERA_SCCB_SDA_PIN=15
+CONFIG_BSP_CAMERA_SCCB_SDA0_PIN=0
+CONFIG_BSP_CAMERA_SCCB_SCLK_PIN=14
+CONFIG_BSP_CAMERA_CMOS_RST_PIN=13
+CONFIG_BSP_CAMERA_CMOS_VSYNC_PIN=12
+CONFIG_BSP_CAMERA_CMOS_PWDN_PIN=11
+CONFIG_BSP_CAMERA_CMOS_XCLK_PIN=9
+CONFIG_BSP_CAMERA_CMOS_PCLK_PIN=8
+CONFIG_BSP_CAMERA_CMOS_HREF_PIN=10
 # CONFIG_BSP_USING_RW007 is not set
 # CONFIG_BSP_USING_RTC is not set
 # CONFIG_BSP_USING_WDT is not set
@@ -1050,7 +1062,13 @@ CONFIG_RT_AI_USE_K210=y
 # mpy-extmods
 #
 CONFIG_PRJ_USING_EXTMODS_MISC=y
-# CONFIG_EXTMODS_MISC_USING_K210 is not set
+CONFIG_EXTMODS_MISC_USING_K210=y
+# CONFIG_EXTMODS_K210_LCDCTL is not set
+CONFIG_EXTMODS_K210_DVP=y
+# CONFIG_EXTMODS_K210_I2S is not set
+# CONFIG_EXTMODS_K210_FPIOA is not set
+# CONFIG_EXTMODS_K210_SHA256 is not set
+# CONFIG_EXTMODS_K210_FFT is not set
 # CONFIG_PKG_USING_OPENMV_CP is not set
 CONFIG_PRJ_USING_RT_AK_EXMODS=y
 # end of mpy-extmods
