+++ project_0/.config.old
@@ -572,7 +572,8 @@ CONFIG_PKG_MICROPYTHON_PATH="/packages/language/micropython"
 #
 # System Module
 #
-# CONFIG_MICROPYTHON_USING_UOS is not set
+CONFIG_MICROPYTHON_USING_UOS=y
+CONFIG_MICROPYTHON_USING_FILE_SYNC_VIA_IDE=y
 # CONFIG_MICROPYTHON_USING_THREAD is not set
 # CONFIG_MICROPYTHON_USING_USELECT is not set
 # CONFIG_MICROPYTHON_USING_UCTYPES is not set
