+++ project_0/mpy_rt_ai/qstrdefs_rt_ai.h
@@ -17,3 +17,4 @@ QDEF(MP_QSTR_ai_find, (const byte*)"\x97\x07" "ai_find")
 QDEF(MP_QSTR_ai_run, (const byte*)"\x1b\x06" "ai_run")
 QDEF(MP_QSTR_ai_output, (const byte*)"\x0d\x09" "ai_output")
 QDEF(MP_QSTR_ai_load, (const byte*)"\x34\x07" "ai_load")
+QDEF(MP_QSTR_ai_free, (const byte*)"\xa6\x07" "ai_free")
